feat(tools): 重构工具模块并优化下载功能
- 将工具模块拆分为download和share两个子模块
- 新增视频打包下载功能(downloadVideosAsZip)
- 优化Live图片加载状态处理和UI显示
- 修复抖音视频解析URL直接使用play_addr_h264.url_list
- 登录成功后使用replace导航避免历史记录问题
- 优化UniversalImage组件视频加载和播放逻辑
