+++ packages/web/src/lib/tools/share/index.ts
@@ -1,0 +1,1 @@
+export *  from './share'
